- Moved EmptyComponent into components.
